+++ lib/philiprehberger/structured_logger.rb
@@ -2,6 +2,7 @@
 
 require_relative "structured_logger/version"
 require_relative "structured_logger/formatter"
+require_relative "structured_logger/async_writer"
 require_relative "structured_logger/logger"
 
 module Philiprehberger